feat:课程计划修改接口

+++ xuecheng-plus-content/src/main/java/com/suave/content/entity/Teachplan.java
@@ -54,7 +54,7 @@ public class Teachplan implements Serializable {
     private String timelength;
 
     @Schema(title = "排序字段")
-    private Integer orderby;
+    private Long orderby;
 
     @Schema(title = "课程标识")
     private Long courseId;
